10 - [Tech] Class managementq - move Resource.kt and Status.kt to domain.common dir
